Refactor package name to be compliant with poetry reqs
