+++ mrcepid_test_loader/loader.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 
 from general_utilities.import_utils.module_loader.module_loader import ModuleLoader
-from test_loader import test_loader_ingester
-from test_loader.test_loader_association_pack import TestLoaderProgramArgs, TestLoaderAssociationPack
+from mrcepid_test_loader import test_loader_ingester
+from mrcepid_test_loader.test_loader_association_pack import TestLoaderProgramArgs, TestLoaderAssociationPack
 
 
 class LoadModule(ModuleLoader):
